Move all storybook stories to /.storybook dir
