+++ .storybook/main.js
@@ -1,10 +1,7 @@
 const customWebpackConfig = require('../webpack.config.js');
 
 module.exports = {
-  stories: [
-    '../src/stories/*.jsx',
-    '../src/stories/gen3-ui-component/index.js',
-  ],
+  stories: ['./stories/*.jsx', './stories/gen3-ui-component/index.js'],
   addons: ['@storybook/addon-actions'],
   webpackFinal: (config) => {
     return {
